renamed package. refactored travis and setup to reflect the change.
