+++ .idea/DiTTo_YoutubePredictor.iml
@@ -9,7 +9,7 @@
     <exclude-output />
     <content url="file://$MODULE_DIR$">
       <sourceFolder url="file://$MODULE_DIR$/tests" isTestSource="true" />
-      <sourceFolder url="file://$MODULE_DIR$/ypTester" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/youtubePredictor" isTestSource="false" />
     </content>
     <orderEntry type="jdk" jdkName="Python 3.9" jdkType="Python SDK" />
     <orderEntry type="sourceFolder" forTests="false" />
