+++ .idea/workspace.xml
@@ -1,7 +1,34 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="ChangeListManager">
-    <list default="true" id="b1b57f69-993d-402a-a7bd-bfabe35b29c6" name="Default Changelist" comment="reorg" />
+    <list default="true" id="b1b57f69-993d-402a-a7bd-bfabe35b29c6" name="Default Changelist" comment="reorg">
+      <change beforePath="$PROJECT_DIR$/.idea/DiTTo_YoutubePredictor.iml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/DiTTo_YoutubePredictor.iml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/setup.py" beforeDir="false" afterPath="$PROJECT_DIR$/setup.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/tests/youtubePredictor_app_testRunner.py" beforeDir="false" afterPath="$PROJECT_DIR$/tests/youtubePredictor_app_testRunner.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/tests/youtubePredictor_app_testSuite.py" beforeDir="false" afterPath="$PROJECT_DIR$/tests/youtubePredictor_app_testSuite.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/tests/youtubePredictor_dataManager_testSuite.py" beforeDir="false" afterPath="$PROJECT_DIR$/tests/youtubePredictor_dataManager_testSuite.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/ypTester/__init__.py" beforeDir="false" afterPath="$PROJECT_DIR$/youtubePredictor/__init__.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/ypTester/static/styles/style.css" beforeDir="false" afterPath="$PROJECT_DIR$/youtubePredictor/static/styles/style.css" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/ypTester/templates/youtubePredictor_administrator.html" beforeDir="false" afterPath="$PROJECT_DIR$/youtubePredictor/templates/youtubePredictor_administrator.html" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/ypTester/templates/youtubePredictor_data.html" beforeDir="false" afterPath="$PROJECT_DIR$/youtubePredictor/templates/youtubePredictor_data.html" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/ypTester/templates/youtubePredictor_emptyDatabase.html" beforeDir="false" afterPath="$PROJECT_DIR$/youtubePredictor/templates/youtubePredictor_emptyDatabase.html" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/ypTester/templates/youtubePredictor_exportDatabase.html" beforeDir="false" afterPath="$PROJECT_DIR$/youtubePredictor/templates/youtubePredictor_exportDatabase.html" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/ypTester/templates/youtubePredictor_save.html" beforeDir="false" afterPath="$PROJECT_DIR$/youtubePredictor/templates/youtubePredictor_save.html" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/ypTester/templates/youtubePredictor_submissionSuccess.html" beforeDir="false" afterPath="$PROJECT_DIR$/youtubePredictor/templates/youtubePredictor_submissionSuccess.html" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/ypTester/templates/youtubePredictor_updateDatabase.html" beforeDir="false" afterPath="$PROJECT_DIR$/youtubePredictor/templates/youtubePredictor_updateDatabase.html" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/ypTester/templates/youtubePredictor_user.html" beforeDir="false" afterPath="$PROJECT_DIR$/youtubePredictor/templates/youtubePredictor_user.html" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/ypTester/templates/youtubePredictor_videoSubmission.html" beforeDir="false" afterPath="$PROJECT_DIR$/youtubePredictor/templates/youtubePredictor_videoSubmission.html" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/ypTester/templates/youtubePredictor_webStreamer.html" beforeDir="false" afterPath="$PROJECT_DIR$/youtubePredictor/templates/youtubePredictor_webStreamer.html" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/ypTester/templates/youtubePredictory_batch.html" beforeDir="false" afterPath="$PROJECT_DIR$/youtubePredictor/templates/youtubePredictory_batch.html" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/ypTester/youtubePredictor_app.py" beforeDir="false" afterPath="$PROJECT_DIR$/youtubePredictor/youtubePredictor_app.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/ypTester/youtubePredictor_constants.py" beforeDir="false" afterPath="$PROJECT_DIR$/youtubePredictor/youtubePredictor_constants.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/ypTester/youtubePredictor_dataManager.py" beforeDir="false" afterPath="$PROJECT_DIR$/youtubePredictor/youtubePredictor_dataManager.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/ypTester/youtubePredictor_databaseBuilder.py" beforeDir="false" afterPath="$PROJECT_DIR$/youtubePredictor/youtubePredictor_databaseBuilder.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/ypTester/youtubePredictor_forms.py" beforeDir="false" afterPath="$PROJECT_DIR$/youtubePredictor/youtubePredictor_forms.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/ypTester/youtubePredictor_logger.py" beforeDir="false" afterPath="$PROJECT_DIR$/youtubePredictor/youtubePredictor_logger.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/ypTester/youtubePredictor_webStreamer.py" beforeDir="false" afterPath="$PROJECT_DIR$/youtubePredictor/youtubePredictor_webStreamer.py" afterDir="false" />
+    </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
     <option name="HIGHLIGHT_NON_ACTIVE_CHANGELIST" value="false" />
@@ -329,7 +356,14 @@
       <option name="project" value="LOCAL" />
       <updated>1616548988468</updated>
     </task>
-    <option name="localTasksCounter" value="18" />
+    <task id="LOCAL-00018" summary="added TODO items">
+      <created>1616549402659</created>
+      <option name="number" value="00018" />
+      <option name="presentableId" value="LOCAL-00018" />
+      <option name="project" value="LOCAL" />
+      <updated>1616549402659</updated>
+    </task>
+    <option name="localTasksCounter" value="19" />
     <servers />
   </component>
   <component name="TypeScriptGeneratedFilesManager">
@@ -360,7 +394,8 @@
     <MESSAGE value="expanded backend test framwork and identified key TODO items" />
     <MESSAGE value="corrected minor typos in logger class" />
     <MESSAGE value="refactoring testing to remediate errors" />
-    <option name="LAST_COMMIT_MESSAGE" value="refactoring testing to remediate errors" />
+    <MESSAGE value="added TODO items" />
+    <option name="LAST_COMMIT_MESSAGE" value="added TODO items" />
   </component>
   <component name="XSLT-Support.FileAssociations.UIState">
     <expand />
