+++ youtubePredictor/static/styles/style.css
