+++ youtubePredictor/templates/youtubePredictor_administrator.html
