+++ youtubePredictor/templates/youtubePredictor_emptyDatabase.html
