+++ youtubePredictor/templates/youtubePredictor_submissionSuccess.html
