+++ youtubePredictor/youtubePredictor_logger.py
@@ -15,7 +15,7 @@
 import logging
 
 # Import DiTTo_YoutubePredictor Utilities
-from ypTester import youtubePredictor_constants as ypConstants
+from youtubePredictor import youtubePredictor_constants as ypConstants
 
 logging.basicConfig(filename="../tests/logs/youtubePredictor.log",
                     format=ypConstants.LOGGING_FORMAT,
